Refactor image sizes and labels in documents.vue

+++ frontend/pages/cities.vue
@@ -1,19 +1,19 @@
 <template>
-    <div class="max-w-5xl mx-auto">
-      <UCard class="my-2">
-        <template #header>
-          <div class="flex justify-between items-center">
-            <h2 class="font-bold">Ciudades</h2>
-            <div class="flex gap-3 my-3">
-                <UInput v-model="search" placeholder="Buscar" />
-                <UPagination v-model="pagination.page" :page-count="pagination.pageSize" :total="pagination.resultsCount" />
-            </div>
+  <div class="max-w-5xl mx-auto">
+    <UCard class="my-2">
+      <template #header>
+        <div class="flex justify-between items-center">
+          <h2 class="font-bold">Ciudades</h2>
+          <div class="flex gap-3 my-3">
+            <UInput v-model="search" placeholder="Buscar" />
+            <UPagination v-model="pagination.page" :page-count="pagination.pageSize" :total="pagination.resultsCount" />
           </div>
-        </template>
-        <div class="flex justify-center items-center">
-          <h3 v-if="cities.length === 0">No hay ciudades</h3>
-        </div>   
-        <div style="overflow: auto;">     
+        </div>
+      </template>
+      <div class="flex justify-center items-center">
+        <h3 v-if="cities.length === 0">No hay ciudades</h3>
+      </div>
+      <div style="overflow: auto;">
         <table class="table-auto w-full permission-table">
           <thead>
             <tr>
@@ -23,25 +23,31 @@
               <th :class="ui.th">Acciones</th>
             </tr>
           </thead>
-          <tbody>            
-            
+          <tbody>
+
             <tr v-for="(citie, index) in cities" :key="index">
               <td :class="ui.td">
                 <div class="flex items-center justify-center">
-                  <UInput v-model="citie.name" @blur="saveItem(index,'name',citie.name)" class="border rounded p-1 " />
-                  <UInput v-model="citie.municipality_dane_code" @blur="saveItem(index,'municipality_dane_code',citie.municipality_dane_code)" class="border rounded p-1 w-14" />
+                  <UInput v-model="citie.name" @blur="saveItem(index, 'name', citie.name)" class="border rounded p-1 " />
+                  <UInput v-model="citie.municipality_dane_code"
+                    @blur="saveItem(index, 'municipality_dane_code', citie.municipality_dane_code)"
+                    class="border rounded p-1 w-14" />
                 </div>
               </td>
               <td :class="ui.td">
                 <div class="flex items-center justify-center">
-                  <UInput v-model="citie.departament" @blur="saveItem(index,'departament',citie.departament)" class="border rounded p-1" />
+                  <UInput v-model="citie.departament" @blur="saveItem(index, 'departament', citie.departament)"
+                    class="border rounded p-1" />
 
-                  <UInput v-model="citie.departament_dane_code" @blur="saveItem(index,'departament_dane_code',citie.departament_dane_code)" class="border rounded p-1 w-10" />
+                  <UInput v-model="citie.departament_dane_code"
+                    @blur="saveItem(index, 'departament_dane_code', citie.departament_dane_code)"
+                    class="border rounded p-1 w-10" />
                 </div>
               </td>
               <td :class="ui.td">
                 <div class="flex items-center justify-center">
-                  <UInput v-model="citie.region" @blur="saveItem(index,'region',citie.region)" class="border rounded p-1 w-auto" />
+                  <UInput v-model="citie.region" @blur="saveItem(index, 'region', citie.region)"
+                    class="border rounded p-1 w-auto" />
                 </div>
 
               </td>
@@ -50,8 +56,8 @@
                   <span @click="deleteCitie(citie.id)" :class="ui.span">🗑️</span>
                 </div>
               </td>
-            </tr> 
-           
+            </tr>
+
             <tr>
               <td :class="ui.td">
                 <div class="flex items-center justify-center">
@@ -77,9 +83,9 @@
           </tbody>
         </table>
       </div>
-      </UCard>   
-    </div>
-  </template>
+    </UCard>
+  </div>
+</template>
 
 
 <script setup lang="ts">
@@ -93,37 +99,48 @@ const newCitieRegion = ref('')
 
 
 const {
-    data: cities ,
-    pagination,
-    search ,
-    pending,
+  data: cities,
+  pagination,
+  search,
+  pending,
 } = usePaginatedFetch<any>("/api/cities/");
 
 const fetchCities = async () => {
   const {
-    data: cities ,
+    data: cities,
     pagination,
-    search ,
+    search,
     pending,
   } = usePaginatedFetch<any>("/api/cities/");
 
-  console .log('fetchCities',cities.value)
-  
+  console.log('fetchCities', cities.value)
+
 }
 
 const deleteCitie = async (id: number) => {
-    const message = confirm('¿Estás seguro de eliminar esta ciudad?')
-    if (message) {
-        const response = await $fetch(`api/cities/${id}/`, {
-            method: 'DELETE'
-        })
-        fetchCities()
-    }
+  autorized.value = await validatePermissions('cities', 'delete');
+  if (!autorized.value) {
+    fetchCities()
+    return
+  }
+  const message = confirm('¿Estás seguro de eliminar esta ciudad?')
+  if (message) {
+    const response = await $fetch(`api/cities/${id}/`, {
+      method: 'DELETE'
+    })
+    fetchCities()
+  }
 }
 
-
+const autorized = ref(false)
 
 const saveItem = async (index: number, field: string, value: string) => {
+  autorized.value = await validatePermissions('cities', 'change');
+  if (!autorized) {   
+    fetchCities()
+    return
+  }
+
   const citie = cities.value[index];
   citie[field] = value;
   const response = await $fetch(`api/cities/${citie.id}`, {
@@ -132,43 +149,48 @@ const saveItem = async (index: number, field: string, value: string) => {
       [field]: value,
     }),
   });
-    fetchCities();
+  fetchCities();
 };
 
 const createCitie = async () => {
-    const message = confirm('¿Estás seguro de crear esta ciudad?')
-
-    if (!message){
-        newCitieName.value = ''
-        newCitieDepartament.value = ''
-        newCitieRegion.value = ''
-        fetchCities()
-        return
-    }
-    
-    const response = await $fetch('api/cities/', {
-        method: 'POST',
-        body: {
-            name: newCitieName.value,
-            departament: newCitieDepartament.value,
-            region: newCitieRegion.value
-        }
-    })
+  autorized.value = await validatePermissions('cities', 'add');
+  if (!autorized) {
     fetchCities()
+    return
+  }
+  const message = confirm('¿Estás seguro de crear esta ciudad?')
+
+  if (!message) {
     newCitieName.value = ''
     newCitieDepartament.value = ''
     newCitieRegion.value = ''
+    fetchCities()
+    return
+  }
+
+  const response = await $fetch('api/cities/', {
+    method: 'POST',
+    body: {
+      name: newCitieName.value,
+      departament: newCitieDepartament.value,
+      region: newCitieRegion.value
+    }
+  })
+  fetchCities()
+  newCitieName.value = ''
+  newCitieDepartament.value = ''
+  newCitieRegion.value = ''
 }
 
 onMounted(() => {
-   fetchCities()
+  fetchCities()
 })
 
 const ui = {
-    td: 'p-1 border',
-    th: 'p-2 border',
-    check: 'align-center justify-center',
-    span: 'cursor-pointer'
+  td: 'p-1 border',
+  th: 'p-2 border',
+  check: 'align-center justify-center',
+  span: 'cursor-pointer'
 }
 
 </script>
